Updated the profile.diet

+++ lib/login.dart
@@ -22,35 +22,35 @@ class _LoginPageState extends State<LoginPage> {
   String _username, _password, _token;
 
   Future<bool> _pullUserDataFromServer(BackendRequest backend) async {
-
     print("Pulling user data from server");
-
     UserProfile profile = await backend.getUserProfile();
     DB.DatabaseHelper database = DB.DatabaseHelper.instance;
-
     // Load diet fresh
-    LocalStorage.deleteDiet();
-    LocalStorage.storeDiet(profile.diet.id);
-
-    print("Loaded diet ${profile.diet.id}");
-
+    if(profile.diet != null) {
+      LocalStorage.deleteDiet();
+      LocalStorage.storeDiet(profile.diet.id);
+      print("Loaded diet ${profile.diet.id}");
+    }
+    
     // Load allergens fresh
-    database.clearAllergens();
-    for(Map<String, dynamic> allergen in profile.allergens) {
-      database.insertAllergen(DB.Allergen(id: allergen['id'], name: allergen['name']));
-      print("Loaded allergen ${allergen['id']}");
+    if(profile.allergens != null) {
+      database.clearAllergens();
+      for(Map<String, dynamic> allergen in profile.allergens) {
+        database.insertAllergen(DB.Allergen(id: allergen['id'], name: allergen['name']));
+        print("Loaded allergen ${allergen['id']}");
+      }
     }
-
     // Load favorites fresh
-    database.clearRecipes();
-    for(Map<String, dynamic> recipe in profile.favorites) {
-      database.insertRecipe(DB.Recipe(id: recipe['api_id'], name: "n/a", img: "n/a"));
-      print("Loaded favorite recipe ${recipe['api_id']}");
-
+    if(profile.favorites != null) {
+      database.clearRecipes();
+      for(Map<String, dynamic> recipe in profile.favorites) {
+        database.insertRecipe(DB.Recipe(id: recipe['api_id'], name: "n/a", img: "n/a"));
+        print("Loaded favorite recipe ${recipe['api_id']}");
+      }
     }
-
     return true;
-  }
+}
+  
 
   _submit() async {
     if (_formKey.currentState.validate()) {
